+++ index.html
@@ -31,6 +31,8 @@
             </div>
         </div>
     </section>
-    <script src="./script/main.js"></script>
+    <script src="./script/firstInput.js" type="module"></script>
+    <script src="./script/addTask.js" type="module"></script>
+    <script src="./script/main.js" type="module"></script>
 </body>
 </html>
